Start UI implemented
use PuF\src\gameUi\Start.java to start the App
