--- src/application/View/View.java
@@ -1,7 +1,0 @@
-package application.View;
-
-public class View {
-
-
-
-}
